Added "WaitList" class to wait functions.  Updates.
Still working on rewriting the waiting code.  There are compile errors
in this file, but it's not included in the project, so I'm not too
worried about it.

I updated the wait_functions to use a WaitList class (that will be a
Singleton) and to use a std::list implementation instead of the custom
implementation of a linked list that is currently being used.  This
makes the code much easier to read.

I still need to bring some functions over and/or create another light
utility class that incorporates the missing functionality.