Add helper code.
Adding classes that will be used to write the combat manager
(eventually).  Extensive testing will need to be done to ensure that
damage is more-or-less the same for combat before and after the changes.

The combat manager doesn't exist yet, so this is a safe check-in.

I also added in some getter functions to the obj_data structure and the
obj_flags structure to make it easier to access nested data.  I should
really do this for char_data.

This changelist also includes some additional work on my WaitList
implementation that I have taken a break from.