Recover and Beats
Recover now will check if the character is wearing a quiver on their
back and put the arrows back in the quiver. If the character doesn't
have a quiver they are just placed in their inventory.

The beats on archery has been reduced to better reflect the average
character. Beats_modifier has been replaced with just the base_beats.
Also, I added a check to make sure the beats are never longer than 12.