Big Brother work.
When the last item is removed from a corpse, that player is no longer
considered looting.

If someone gets attacked in PK, mark them as a PKer for the sake of not
getting AFK protection.

When a player dies, mark them as looting.

Use IDs to store the looting set instead of pointers.