Archery Spec
Added a set shooting command and only allowed it for the archery spec.
Shooting fast will cut the beats in half and the damage. Shooting slow
will increase beats times 2 and damage. Also added the framework for set
casting. This will probably be an arcane specialization for those
players that don't want to select a favorite element.