Pre PK Test Run
Added two new guildmasters for ALT rots and the release of the new
specs. Reduced the damage of ambush unless the player is spec in stealth
after level 20. Changed spell save to use a 1d8 instead of 1d20 because
mobs and players were saving every other casting. Also allow the casting
of lightning bolt inside if you are specd in lightning.