Fix for Riposte with Bears and Fame
Riposte used a body part of 21 which was out of the scope for Beornings.
I change this to a static body part as a quick fix, but it needs to be
set to a random body part selection.

I had to revert the entire pkill.c to it's original state because if the
fame file was empty it would crash the game when trying to write to it.
This needs to be overhauled to fix the underlining issue.