Fighter spec iteration.
Move wild fighting and weapon master into their own structs that contain all of the logic of the spec.  These should probably live in a fighter_specs.h/cpp file in the future.

Buff weapon master spec to be more competitive.

Add score line to weapon master spec that's weapon specific.

Refactor code to use new structures.  This REALLY shows the need for an event system or an ECS implementation.